feat: add ticket validation functionality and update session deletion logic

+++ src/services/sessionService.ts
@@ -20,11 +20,10 @@ export class SessionService {
   }
 
   static async delete(sessionId: number) {
+    const session = await prisma.session.findUnique({ where: { id: sessionId } });
+    if(!session) return;
+    
     return prisma.session.delete({ where: { id: sessionId } });
-  }
-
-  static async deleteByRefreshToken(refreshToken: string) {
-    return prisma.session.deleteMany({ where: { refreshToken } });
   }
 
   static async deleteByUserId(userId: number) {
